+++ src/test/java/ForLoopTest/NewTest.java
@@ -1,3 +1,5 @@
+package ForLoopTest;
+
 import org.testng.annotations.AfterTest;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
