+++ src/test/java/ForLoopTest/trycatchtest.java
@@ -1,3 +1,5 @@
+package ForLoopTest;
+
 import org.testng.annotations.Test;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.AfterTest;
