+++ src/test/java/TheInternet/foundation/ConfigurationReader.java
@@ -1,3 +1,5 @@
+package TheInternet.foundation;
+
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
