+++ src/test/java/TheInternet/foundation/PageObjectBase.java
@@ -1,9 +1,11 @@
+package TheInternet.foundation;
+
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.PageFactory;
 
 public abstract class PageObjectBase {
 	protected WebDriver driver;
-	private String url;
+	protected String url;
 	
 	public PageObjectBase(WebDriver driver, String url) {
 		this.driver = driver;
