+++ src/test/java/TheInternet/pages/DropdownPage.java
@@ -1,8 +1,11 @@
+package TheInternet.pages;
+
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 import org.openqa.selenium.support.ui.Select;
+import TheInternet.foundation.PageObjectBase;
 
 public class DropdownPage extends PageObjectBase {
