+++ src/test/java/TheInternet/pages/IndexPage.java
@@ -1,3 +1,5 @@
+package TheInternet.pages;
+
 import org.openqa.selenium.WebDriver;
 
 public class IndexPage {
