+++ src/test/java/TheInternet/tests/SeleniumTest.java
@@ -1,3 +1,5 @@
+package TheInternet.tests;
+
 import org.testng.annotations.Test;
 import org.testng.annotations.BeforeTest;
 import org.openqa.selenium.WebDriver;
